GS: Split GS.h into multiple smaller headers
