+++ pcsx2/GS/GS.cpp
@@ -17,7 +17,7 @@
 #include "GS/Window/GSwxDialog.h"
 #include "GS.h"
 #include "GSUtil.h"
-#include "GS_types.h"
+#include "GSExtra.h"
 #include "Renderers/SW/GSRendererSW.h"
 #include "Renderers/Null/GSRendererNull.h"
 #include "Renderers/Null/GSDeviceNull.h"
