+++ pcsx2/GS/GSDrawingContext.cpp
@@ -15,6 +15,7 @@
 
 #include "PrecompiledHeader.h"
 #include "GSDrawingContext.h"
+#include "GSGL.h"
 #include "GS.h"
 
 static int findmax(int tl, int br, int limit, int wm, int minuv, int maxuv)
