+++ pcsx2/GS/GSDrawingEnvironment.h
@@ -15,8 +15,6 @@
 
 #pragma once
 
-#include "GS.h"
-
 class alignas(32) GSDrawingEnvironment
 {
 public:
