+++ pcsx2/GS/GSLocalMemory.h
@@ -15,7 +15,6 @@
 
 #pragma once
 
-#include "GS.h"
 #include "GSTables.h"
 #include "GSVector.h"
 #include "GSBlock.h"
