+++ pcsx2/GS/GSPng.cpp
@@ -15,7 +15,7 @@
 
 #include "PrecompiledHeader.h"
 #include "GSPng.h"
-#include "GS_types.h"
+#include "GSExtra.h"
 #include <zlib.h>
 #include <png.h>
