+++ pcsx2/GS/GSUtil.cpp
@@ -14,6 +14,8 @@
  */
 
 #include "PrecompiledHeader.h"
+#include "GS.h"
+#include "GSExtra.h"
 #include "GSUtil.h"
 #include <locale>
 #include <codecvt>
