+++ pcsx2/GS/GSVector.h
@@ -14,6 +14,7 @@
  */
 
 #include "PrecompiledHeader.h"
+#include "GSIntrin.h"
 
 #pragma once
