+++ pcsx2/GS/Renderers/Common/GSOsdManager.h
@@ -21,6 +21,7 @@
 #include "GSTexture.h"
 #include "ft2build.h"
 #include FT_FREETYPE_H
+#include <map>
 
 class GSOsdManager
 {
