+++ pcsx2/GS/Renderers/Common/GSRenderer.cpp
@@ -15,8 +15,8 @@
 
 #include "PrecompiledHeader.h"
 #include "GSRenderer.h"
-#include "Host.h"
-#include "pcsx2/Config.h"
+#include "gui/AppConfig.h"
+#include "GS/GSGL.h"
 #if defined(__unix__)
 #include <X11/keysym.h>
 #endif
