+++ pcsx2/GS/Renderers/Common/GSRenderer.h
@@ -15,7 +15,6 @@
 
 #pragma once
 
-#include "GS/GS.h"
 #include "GS/GSState.h"
 #include "GS/GSCapture.h"
