+++ pcsx2/GS/Renderers/DX11/GSDevice11.cpp
@@ -17,6 +17,7 @@
 #include "GS.h"
 #include "GSDevice11.h"
 #include "GS/Renderers/DX11/D3D.h"
+#include "GS/GSExtra.h"
 #include "GS/GSUtil.h"
 #include "GS/resource.h"
 #include <fstream>
