+++ pcsx2/GS/Renderers/HW/GSHwHack.cpp
@@ -15,7 +15,6 @@
 
 #include "PrecompiledHeader.h"
 #include "GS/GSState.h"
-#include "GS.h"
 
 bool s_nativeres;
 static CRCHackLevel s_crc_hack_level = CRCHackLevel::Full;
