+++ pcsx2/GS/Renderers/HW/GSRendererHW.cpp
@@ -15,6 +15,7 @@
 
 #include "PrecompiledHeader.h"
 #include "GSRendererHW.h"
+#include "GS/GSGL.h"
 
 const float GSRendererHW::SSR_UV_TOLERANCE = 1e-3f;
