+++ pcsx2/GS/Renderers/OpenGL/GLLoader.cpp
@@ -15,7 +15,7 @@
 
 #include "PrecompiledHeader.h"
 #include "GLLoader.h"
-#include "GS.h"
+#include "GS/GS.h"
 #include <unordered_set>
 
 namespace GLExtension
