+++ pcsx2/GS/Renderers/OpenGL/GSDeviceOGL.h
@@ -19,7 +19,6 @@
 #include "common/GL/StreamBuffer.h"
 #include "GS/Renderers/Common/GSDevice.h"
 #include "GSTextureOGL.h"
-#include "GS/GS.h"
 #include "GSUniformBufferOGL.h"
 #include "GSShaderOGL.h"
 #include "GLState.h"
