+++ pcsx2/GS/Renderers/SW/GSSetupPrimCodeGenerator.all.cpp
@@ -14,7 +14,6 @@
  */
 
 #include "PrecompiledHeader.h"
-#include "GS/GS_types.h"
 #include "GSSetupPrimCodeGenerator.all.h"
 #include "GSVertexSW.h"
