+++ pcsx2/GS/Window/GSCaptureDlg.cpp
@@ -16,7 +16,7 @@
 #include "PrecompiledHeader.h"
 #include "GS.h"
 #include "GSCaptureDlg.h"
-#include "GS/GS_types.h"
+#include "GS/GSExtra.h"
 
 // Ideally this belongs in WIL, but CAUUID is used by a *single* COM function in WinAPI.
 // That's presumably why it's omitted and is unlikely to make it to upstream WIL.
