+++ pcsx2/Gif_Unit.cpp
@@ -16,7 +16,6 @@
 #include "PrecompiledHeader.h"
 #include "Common.h"
 
-#include "GS.h"
 #include "Gif_Unit.h"
 #include "Vif_Dma.h"
 #include "MTVU.h"
